[workers-auth] Add @cloudflare/workers-auth/cf entrypoint over a shared product-agnostic core
Refactor the wrangler-specific auth layer into a product-agnostic
`createCloudflareAuth(product, ctx)` core (`src/product/`) parameterised
by an `AuthProduct` descriptor (identity, config paths, file format,
consent pages, client id, copy). Wrangler becomes a thin descriptor
(`WRANGLER_PRODUCT` + `createWranglerAuth`) with its `/wrangler`
entrypoint exports unchanged.

Add a new `@cloudflare/workers-auth/cf` entrypoint (`createCfAuth`):
the same auth machinery, but JSON files under `~/.config/cloudflare`,
the `cloudflare` OS-keyring service, and `CLOUDFLARE_CLIENT_ID`. The
on-disk format is a new `FileFormat` axis threaded through the credential
store (default TOML, so wrangler is byte-for-byte unchanged).

cf's OAuth app client-id UUIDs, redirect URI, consent page URLs, and the
account-selection config-file hint are placeholders pending real values
(search TODO(cf)).
