+++ packages/workers-auth/package.json
@@ -26,6 +26,10 @@
 			"import": "./dist/wrangler/index.mjs",
 			"types": "./dist/wrangler/index.d.mts"
 		},
+		"./cf": {
+			"import": "./dist/cf/index.mjs",
+			"types": "./dist/cf/index.d.mts"
+		},
 		"./test-helpers": {
 			"import": "./dist/test-helpers/index.mjs",
 			"types": "./dist/test-helpers/index.d.mts"
